Former-commit-id: 05570e961e936246b33033df0d0a468436d4ac4a
Former-commit-id: b385bd08fa5e6c341edfb5771905b7fcc3c0282d

+++ NINA/PlateSolving/PlateSolveParameter.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 namespace NINA.PlateSolving {
 
     internal class PlateSolveParameter {
-        public int FocalLength { get; set; }
+        public double FocalLength { get; set; }
         public double PixelSize { get; set; }
         public double ImageWidth { get; set; }
         public double ImageHeight { get; set; }
